# php-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invoice Management API project",
        "description": "Invoice Management API project. It provides endpoints to manage users, tax profiles, and invoices. The project is structured using the MVC architecture and includes automated tests, Docker support, and API documentation via Postman.",
        "contact": {
            "email": "caiazza.domenico@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List of users",
                "description": "Returns a paginated collection of users",
                "operationId": "d7a57f4d7de18fb0f92501172472c9b6",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/xApiKeyHeader"
                    },
                    {
                        "name": "filter[email]",
                        "in": "query",
                        "description": "Filter by email",
                        "required": false,
                        "schema": {
                            "type": "steing",
                            "format": "email",
                            "example": "mario@example.com"
                        }
                    },
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Filter by name",
                        "required": false,
                        "schema": {
                            "type": "steing",
                            "example": "Mario"
                        }
                    },
                    {
                        "name": "filter[surname]",
                        "in": "query",
                        "description": "Filter by surname",
                        "required": false,
                        "schema": {
                            "type": "steing",
                            "example": "Rossi"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Links"
                                            }
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Meta"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "description": "Creates a new user after validating the input data and returns the created user's details.",
                "operationId": "9fea6f2adcb69614c521a7733358382d",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/xApiKeyHeader"
                    }
                ],
                "requestBody": {
                    "description": "The required data for creating a new user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "name",
                                    "surname",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntity"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get a single user",
                "description": "Returns a single user by their ID",
                "operationId": "40a5104e23b52137e78c3d3ac48b88a8",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/xApiKeyHeader"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User getted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update an existing user",
                "description": "Updates an existing user's data. Only the provided fields will be updated.",
                "operationId": "b166772cec9714067040e77cf14beeba",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/xApiKeyHeader"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data for updating the user. Include only the fields you wish to change.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newSecret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User updated successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No content, meaning no fields were changed"
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user",
                "description": "Deletes a user by their ID and returns no content.",
                "operationId": "ce06860a51ab016ef66e753544fe4563",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/xApiKeyHeader"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted successfully, no content returned."
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Links": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "[url]?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "[url]?page=1"
                    },
                    "prev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Meta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "type": "string",
                                    "example": null,
                                    "nullable": true
                                },
                                "label": {
                                    "type": "string",
                                    "example": "&laquo; Previous"
                                },
                                "active": {
                                    "type": "boolean",
                                    "example": false
                                }
                            },
                            "type": "object"
                        }
                    },
                    "path": {
                        "type": "string",
                        "example": "http://[path]"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "to": {
                        "type": "integer",
                        "example": 2
                    },
                    "total": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "NotFound": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Record not found"
                    }
                },
                "type": "object"
            },
...
            "UnprocessableEntity": {
                "properties": {
                    "message": {
                        "properties": {
                            "Field name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Field error message"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "User": {
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Mario"
                    },
                    "surname": {
                        "type": "string",
                        "example": "Rossi"
                    },
                    "email": {
                        "type": "string",
                        "example": "mario@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-01-01 00:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-01-01 00:00:00"
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "per_page": {
                "name": "per_page",
                "in": "query",
                "description": "Results per page",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 10
                }
            },
            "page": {
                "name": "page",
                "in": "query",
                "description": "Results page",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 1
                }
            },
            "xApiKeyHeader": {
                "name": "x-api-key",
                "in": "header",
                "description": "API key authentication",
                "required": true,
                "schema": {
                    "type": "string",
                    "example": "secret"
                }
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "x-api-key",
                "in": "header"
            }
        }
    },
...
}